refactor: split parse_jest_fn_call
